fixed compiler warnings

+++ test_tf2/test/test_message_filter.cpp
@@ -311,7 +311,7 @@ TEST(MessageFilter, tolerance)
   EXPECT_EQ(0, n.count_); //No return due to lack of space for offset
 
   double time_stamp = (stamp.sec + stamp.nanosec/1e9) + tf2::durationToSec(offset)*1.1;
-  builtin_interfaces::msg::Time stamp_transform = rclcpp::Time((int)time_stamp, (time_stamp - (int)time_stamp)*1e9);
+  builtin_interfaces::msg::Time stamp_transform = rclcpp::Time(static_cast<int32_t>((int)time_stamp), static_cast<uint32_t>((time_stamp - (int)time_stamp)*1e9));
 
   buffer.setTransform(createTransform(tf2::Quaternion(0,0,0,1), tf2::Vector3(1,2,3), stamp_transform, "frame1", "frame2"), "me");
 
@@ -319,7 +319,7 @@ TEST(MessageFilter, tolerance)
 
   time_stamp = (stamp.sec + stamp.nanosec/1e9) + tf2::durationToSec(offset);
 
-  msg->header.stamp = rclcpp::Time(time_stamp);
+  msg->header.stamp = rclcpp::Time(static_cast<int64_t>(time_stamp));
   filter.add(msg);
 
   EXPECT_EQ(1, n.count_); // Latest message is off the end of the offset
